runtime: own the Main display control adapter
