+++ docs/agents/architecture-debt-ledger.md
@@ -41,7 +41,7 @@ or record an explicit behavior change.
 |---|---|
 | Main command transport | `mister/platform/runtime/src/main_command.rs`; launcher callers use typed `MainCommand` variants and have no endpoint access |
 | Launch handoff and recovery | `execute_game_launch_with` now builds the portable `LaunchHandoffRequest` and invokes `LaunchIoHandoff`; `SystemLaunchIo` remains the temporary production effect adapter until its raw process, marker, profile, override, and Main calls move behind their owning runtime/persistence capabilities. Recovery remains in `spawn_mister`, `reboot_mister_with`, and `exit_to_mister`. |
-| Display control | `LauncherDisplayControl` implements the portable `DisplayControl` capability over typed `MainCommand` calls; the public display helpers retain their existing strings, polling cadence, and transaction-state API as compatibility wrappers. |
+| Display control | Platform runtime's `MainDisplayControl` implements the portable `DisplayControl` capability over typed `MainCommand` calls and owns response parsing; the public launcher display helpers retain their existing strings, polling cadence, and transaction-state API as compatibility wrappers. |
 | Runtime state and process inspection | Platform runtime's `SystemRuntimeState` implements the portable `RuntimeState` snapshot for registered Main process state, arcade-core command-line classification, and heartbeat; launcher compatibility helpers now delegate to that capability. |
 | Launcher persistence | `SystemLauncherPersistence` implements the portable `LauncherPersistence` capability for launch-return state, settings, input policy, and rebuild-marker ownership; launcher-facing helpers remain compatibility wrappers. Button profiles, screenshot-pack cleanup, and menu-wallpaper restoration remain explicit launcher composition effects. |
